Update README.md
Created a working Soda Machine (or "Pop", here in Michigan) that allows for a customer to purchase a soda. The soda machine will take the payment, produce correct change, and dispense the soda back to the customer. If not enough money is passed in, do not complete the transaction and return the coins back to the customer. Start with a specific inventory for each soda choice, and only dispense the chosen soda if enough of that item remains in stock. 

Project was built using SOLID design priniciples and C# best practices.